+++ Js2IL.Tests/Node/Stream/GeneratorTests.cs
@@ -21,5 +21,17 @@ namespace Js2IL.Tests.Node.Stream
         [Fact]
         public Task Stream_Pipe_Basic()
             => GenerateTest(nameof(Stream_Pipe_Basic));
+
+        [Fact]
+        public Task Stream_PassThrough_Basic()
+            => GenerateTest(nameof(Stream_PassThrough_Basic));
+
+        [Fact]
+        public Task Stream_Transform_Basic()
+            => GenerateTest(nameof(Stream_Transform_Basic));
+
+        [Fact]
+        public Task Stream_Pipe_Backpressure_Basic()
+            => GenerateTest(nameof(Stream_Pipe_Backpressure_Basic));
     }
 }